refactor:minify upload file logic
